+++ tests/test_shell_interactive.py
@@ -5,8 +5,9 @@ import queue
 import unittest
 
 from executorlib import Executor
-from executorlib.shared.executor import cloudpickle_register, execute_parallel_tasks
-from executorlib.shared.spawner import MpiExecSpawner
+from executorlib.standalone.serialize import cloudpickle_register
+from executorlib.interactive.shared import execute_parallel_tasks
+from executorlib.standalone.interactive.spawner import MpiExecSpawner
 
 
 executable_path = os.path.join(os.path.dirname(__file__), "executables", "count.py")
